Rename and reshuffle files
