+++ PlayCover/Views/Settings/UpdateSettings.swift
